Fixed bug with non-inotify file monitoring loop
The loop was triggering forever after first change.
Turns out that stat on directories is buggy and reports current timestamp
of the process if the diretory contains files that has been modified.
So we exclude directories from list of files to be monitored.

Also exclude summary log file just in case.